requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.37.1
gitdb==4.0.10
GitPython==3.1.29
gunicorn==20.1.0
idna==3.4
imgviz==1.2.6
importlib-metadata==5.1.0
ipython==7.34.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
labelme==4.5.9
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.2
termcolor==1.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.1.9
wincertstore==0.2
zipp==3.11.0
